+++ client/components/popover/README.md
@@ -3,10 +3,8 @@ Popover
 
 React components that provide support for modal popovers.
 
-- `Popover` is a component that can be used to show any content
-in a popover.
-- `PopoverMenu` is a component based on `Popover` used to show a menu of
-actions in a popover. It is fully keyboard accessible.
+- `Popover` is a component that can be used to show any content in a popover.
+- `PopoverMenu` is a component based on `Popover` used to show a menu of actions in a popover. It is fully keyboard accessible.
 
 ### Common `Popover`/`PopoverMenu` Properties
 
@@ -45,35 +43,53 @@ The `position` property can be one of the following values:
 #### `ignoreContext`
 
 The `ignoreContext` lets you specify a component that you want to be on the inside clickOutside context. 
-So a context that you want to ignore. In most cases this is not needed but if you want to also have a label 
-that can trigger the opening and closing of the Popover then you need to pass in the label component as a reference.
+So a context that you want to ignore. In most cases this is not needed but if you want to also have a label  that can trigger the opening and closing of the Popover then you need to pass in the label component as a reference.
 
 ### `Popover` Usage
 
-```js
-<button ref="popoverButton" className="button"
-		onClick={ this._onTogglePopover }>Show Popover</button>
-<Popover context={ this.refs && this.refs.popoverButton }
-		isVisible={ this.state.showPopover }
-		onClose={ this._closePopover }
-		className='component__popover'
-		position="top">
+```jsx
+<button
+	ref="popoverButton"
+	className="button"
+	onClick={ this._onTogglePopover }
+>
+	Show Popover
+</button>
+
+<Popover
+	context={ this.refs && this.refs.popoverButton }
+	isVisible={ this.state.showPopover }
+	onClose={ this._closePopover }
+	className='component__popover'
+	position="top"
+>
 	Lorem ipsum dolor sit amet.
 </Popover>
 ```
 
 ### `PopoverMenu` Usage
 
-```js
-<button ref="popoverMenuButton" className="button"
-		onClick={ this._onTogglePopoverMenu }>Show Popover Menu</button>
-<PopoverMenu context={ this.refs && this.refs.popoverMenuButton }
-	 	isVisible={ this.state.showPopoverMenu }
-		onClose={ this._closePopoverMenu }
-		position={ this.state.popoverPosition }>
+```jsx
+<button
+	ref="popoverMenuButton"
+	className="button"
+	onClick={ this._onTogglePopoverMenu }
+>
+	Show Popover Menu
+</button>
+
+<PopoverMenu
+	context={ this.refs && this.refs.popoverMenuButton }
+ 	isVisible={ this.state.showPopoverMenu }
+	onClose={ this._closePopoverMenu }
+	position={ this.state.popoverPosition }
+>
 	<PopoverMenuItem action="A">Item A</PopoverMenuItem>
-	<PopoverMenuItem action="B"
-			onClick={ this._onPopoverMenuItemBClick }>Item B</PopoverMenuItem>
+	<PopoverMenuItem
+			action="B"
+			onClick={ this._onPopoverMenuItemBClick }>
+			Item B
+	</PopoverMenuItem>
 	<PopoverMenuSeparator />
 	<PopoverMenuItem action="C">Item C</PopoverMenuItem>
 </PopoverMenu>
